+++ test/lib/solid_queue_guard/mission_control/navigation_extension_test.rb
@@ -47,6 +47,10 @@ module SolidQueueGuard
           false
         end
 
+        def batches_supported?
+          false
+        end
+
         def jobs_count_with_status(_status)
           0
         end